requirements.txt
fastapi
uvicorn[standard]
gunicorn
boto3
python-jose[cryptography]
httpx
cachetools